添加管理员编辑游戏记录功能

+++ src/main/java/cn/com/nl/evaluation/info/detail/dao/UpdateInfoDao.java
@@ -18,7 +18,7 @@ public interface UpdateInfoDao {
 	 * @param gameInfoId 游戏信息ID
 	 * @return
 	 */
-	public long doSelectModifyCount(String gameInfoId);
+	public int doSelectModifyCount(String gameInfoId);
 
 	/**
 	 *
